+++ backend/Backend/ghc/src/main/java/com/hc/ghc/repositorio/RepositorioMedico.java
@@ -1,9 +1,8 @@
 package com.hc.ghc.repositorio;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.hc.ghc.modelo.Medico;
 
-public interface RepositorioMedico extends JpaRepository< Medico, Integer>{
+public interface RepositorioMedico extends JpaRepository< Medico, Long>{
     
 }
