+++ backend/Backend/ghc/src/main/java/com/hc/ghc/repositorio/RepositorioPaciente.java
@@ -1,8 +1,11 @@
 package com.hc.ghc.repositorio;
 
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.hc.ghc.modelo.Paciente;
 
-public interface RepositorioPaciente extends JpaRepository<Paciente, Integer>{
+public interface RepositorioPaciente extends JpaRepository<Paciente, Long>{
+
 }
