+++ backend/Backend/ghc/src/main/java/com/hc/ghc/repositorio/RepositorioProgramadorDeCitas.java
@@ -1,9 +1,10 @@
 package com.hc.ghc.repositorio;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 
+import org.springframework.data.jpa.repository.JpaRepository;
 import com.hc.ghc.modelo.Programadordecitas;
 
-public interface RepositorioProgramadorDeCitas extends JpaRepository<Programadordecitas, Integer>{
+public interface RepositorioProgramadorDeCitas extends JpaRepository<Programadordecitas, Long>{
+
     
 }
